requirements.txt
pystray
Pillow
pyinstaller